requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
audioread=3.0.1=pypi_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
filelock=3.15.4=pypi_0
fonttools=4.53.0=pypi_0
fsspec=2024.6.1=pypi_0
idna=3.7=pypi_0
importlib-resources=6.4.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.10.2.post1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.43.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.40=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py39h06a4308_0
platformdirs=4.2.2=pypi_0
pooch=1.8.2=pypi_0
pycparser=2.22=pypi_0
pydub=0.25.1=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=69.5.1=py39h06a4308_0
six=1.16.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torch=2.3.1=pypi_0
torchaudio=2.3.1=pypi_0
triton=2.3.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
